diet_recommendation/requirements.txt
absl-py==2.1.0
asgiref==3.8.1
astunparse==1.6.3
cachetools==5.4.0
certifi==2024.7.4
charset-normalizer==3.3.2
diff-match-patch==20230430
Django==5.1
django-import-export==4.1.1
flatbuffers==24.3.25
gast==0.6.0
google-auth==2.33.0
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
grpcio==1.65.5
gunicorn==23.0.0
h5py==3.11.0
idna==3.7
joblib==1.4.2
keras==3.5.0
Keras-Preprocessing==1.1.2
libclang==18.1.1
Markdown==3.7
markdown-it-py==3.0.0
MarkupSafe==2.1.5
mdurl==0.1.2
ml-dtypes==0.4.0
namex==0.0.8
numpy==1.26.4
oauthlib==3.2.2
opt-einsum==3.3.0
optree==0.12.1
packaging==24.1
pandas==2.2.2
plotly==5.23.0
protobuf==3.20.3
pyasn1==0.6.0
pyasn1_modules==0.4.0
Pygments==2.18.0
python-dateutil==2.9.0.post0
python-decouple==3.8
pytz==2024.1
requests==2.32.3
requests-oauthlib==2.0.0
rich==13.7.1
rsa==4.9
scikit-learn==1.5.1
scipy==1.14.0
six==1.16.0
sqlparse==0.5.1
tablib==3.5.0
tenacity==9.0.0
tensorboard==2.17.1
tensorboard-data-server==0.7.2
tensorboard-plugin-wit==1.8.1
tensorflow-cpu==2.17.0
tensorflow-intel==2.17.0
tensorflow-io-gcs-filesystem==0.31.0
termcolor==2.4.0
tf-estimator-nightly==2.8.0.dev2021122109
threadpoolctl==3.5.0
typing_extensions==4.12.2
tzdata==2024.1
urllib3==2.2.2
Werkzeug==3.0.3
wrapt==1.16.0
